+++ src/core/DesktopManager.ts
@@ -1,5 +1,6 @@
 import Meta from "gi://Meta?version=13";
 import Mtk from "gi://Mtk?version=13";
+import Shell from "gi://Shell?version=13";
 
 import type { LayoutManager, Monitor } from "resource:///org/gnome/shell/ui/layout.js";
 
@@ -33,6 +34,7 @@ const TitleBlacklist: RegExp[] = [
 ];
 
 export interface DesktopManagerParams {
+  shell: Shell.Global;
   display: Meta.Display;
   layoutManager: LayoutManager;
   monitorManager: Meta.MonitorManager;
@@ -46,6 +48,7 @@ export interface DesktopManagerParams {
  */
 export default class implements Publisher<DesktopEvent>, GarbageCollector {
   #gc: GarbageCollection;
+  #shell: Shell.Global;
   #display: Meta.Display;
   #layoutManager: LayoutManager;
   #workspaceManager: Meta.WorkspaceManager;
@@ -53,6 +56,7 @@ export default class implements Publisher<DesktopEvent>, GarbageCollector {
   #dispatchCallbacks: DispatchFn<DesktopEvent>[];
 
   constructor({
+    shell,
     display,
     layoutManager,
     monitorManager,
@@ -60,6 +64,7 @@ export default class implements Publisher<DesktopEvent>, GarbageCollector {
     userPreferences,
   }: DesktopManagerParams) {
     this.#gc = new GarbageCollection();
+    this.#shell = shell;
     this.#display = display;
     this.#layoutManager = layoutManager;
     this.#workspaceManager = workspaceManager;
@@ -110,6 +115,33 @@ export default class implements Publisher<DesktopEvent>, GarbageCollector {
   }
 
   /**
+   * The current pointer location as X/Y coordinates.
+   */
+  get pointer(): [x: number, y: number] {
+    const [x, y] = this.#shell.get_pointer();
+
+    return [x, y];
+  }
+
+  /**
+   * The monitor index that the pointer resides on.
+   */
+  get pointerMonitorIdx(): number {
+    const [mouseX, mouseY] = this.#shell.get_pointer();
+
+    for (const monitor of this.#layoutManager.monitors) {
+      if (
+        monitor.x <= mouseX && mouseX <= (monitor.x + monitor.width) &&
+        monitor.y <= mouseY && mouseY <= (monitor.y + monitor.height)
+      ) {
+        return monitor.index;
+      }
+    }
+
+    return 0;
+  }
+
+  /**
    * Moves a window to another monitor, keeping the relative position of the
    * window's top left corner.
    *
@@ -147,8 +179,8 @@ export default class implements Publisher<DesktopEvent>, GarbageCollector {
    * @param selection The selection to be mapped/projected.
    * @param gridSize The reference grid used to divide the monitor’s work area.
    * @param monitorIdx The monitor for which the selection is being mapped.
-   * @param preview Optional. Deducts the user-configured margin ahead of time.
-   *   The margin is usually deducted during the window resize operation.
+   * @param preview Optional. Deducts the user-configured spacing ahead of time.
+   *   The spacing is usually deducted during the window resize operation.
    * @returns The mapped selection.
    */
   selectionToArea(
@@ -437,8 +469,8 @@ export default class implements Publisher<DesktopEvent>, GarbageCollector {
     target.unmaximize(Meta.MaximizeFlags.BOTH);
 
     // All internal calculations fictively operate as if the actual window frame
-    // size would also incorporate the user-defined window margin. Only when a
-    // window is actually moved this margin gets deducted.
+    // size would also incorporate the user-defined window spacing. Only when a
+    // window is actually moved this spacing gets deducted.
     const spacing = this.#userPreferences.getSpacing();
     x += spacing;
     y += spacing;
@@ -486,7 +518,7 @@ export default class implements Publisher<DesktopEvent>, GarbageCollector {
       right = Math.clamp(inset.right, 0, Math.floor(workArea.width / 2)),
       spacing = this.#userPreferences.getSpacing();
 
-    // The fictitious expansion of the workarea by the user-configured margin
+    // The fictitious expansion of the workarea by the user-configured spacing
     // effectively acts as a countermeasure so that windows do always align with
     // the screen edge, i.e., unless the user explicitly configured an inset.
     workArea.x += left - spacing;